Data updates and WorkoutView additions
Removal of target sets in the data model and integration of them into regular sets. This is more straightforward to deal with. Added target set input for workoutView as well as performance stats for the exercise for the workout. The history button changed to open the full stat page for the exercise.